requirements.txt
pdfplumber
python-dateutil
fastapi
uvicorn
pydantic
pytest
regex
